bootcontrol: Add initial support
This patch adds the basic Boot Control HAL implementation that allows
to operate A/B metadata on 'misc' partition. The HAL is based on
Android reference implementation [1] and provides basic functionality
for all interface methods.

This module should be enabled by adding lines to the .mk-file, like:

  PRODUCT_PACKAGES += bootctrl.am57x

Also, A/B system updates has to be enabled according to [2] (including
on bootloader side).

Test: run 'bootctl' on target device

[1] bootable/recovery/boot_control
[2] https://source.android.com/devices/tech/ota/ab/ab_implement

Change-Id: I19852920834eeb3e98e7d6acf06a554e2beaa736
Signed-off-by: Ruslan Trofymenko <ruslan.trofymenko@linaro.org>